fix replace using glob with listfiles
